Deployment patches

+++ infra/provision.sh
@@ -1,5 +1,3 @@
-#!/bin/bash
-
 ###################################################################
 #
 # Provision a Tunnelvision production instance
@@ -11,24 +9,33 @@
 # Close stdin
 exec 0<&-
 
+set -e
+
+rm -rf tunnelvision.zip tunnelvision-*
+
 # Update runtime
 yum -y update
 
+yes | amazon-linux-extras install docker
+service docker start
+
 # Free up port 22 and prevent monkey-business via SSH
-chkconfig sshd off
-service sshd stop
-yum -y erase openssh-server
+if which sshd; then
+  chkconfig sshd off || true
+  service sshd stop || true
+  yum -y erase openssh-server
+fi
 
 # Mount previous Docker volumes if any
-#mount
 
 # Profit
 curl -s https://api.github.com/repos/keksipurkki/tunnelvision/releases/latest \
 | grep "browser_download_url.*zip" \
-| cut -d '"' -f 4 \
-| wget -qi -
+| cut -d '"' -f 4 > latest.url
 
+curl -Lo tunnelvision.zip $(< latest.url)
 unzip tunnelvision.zip
+cd tunnelvision-*/
 
 cat > .env <<EOF
 # AWS
@@ -50,4 +57,5 @@ EMAIL=admin@tunnelvision.me
 
 EOF
 source .env
+docker swarm init
 docker stack deploy $STACK_NAME -c production.yml
